experiment training 4/19/2026 with PSNR 15.39dB
Key fixes for 4D Gaussian temporal parameter stability:

1. Fixed 4D regularization sign - was minimizing instead of maximizing
   - Now returns negative value so minimizing loss = maximizing regularization
   - Encourages high base opacity when temporal opacity is high

2. Fixed coordinate system alignment
   - Set normalize=False in FreeTimeParser to keep cameras in original coordinates
   - NPZ points from triangulation are in the same original system

3. Added temporal parameter stability fixes:
   - Time anchoring regularization (lambda_time_anchor=1e-2)
   - Bidirectional duration regularization (penalize both too-wide AND too-narrow)
   - Hard duration clamping to prevent temporal collapse
   - Time clamping to [0, 1] range

4. Fixed dtype issue in FreeTime_dataset.py (float64 -> float32)

5. Added outlier filtering to combine_frames_fast_keyframes.py
   - Percentile-based filtering removes triangulation outliers

Results on actor1_4_subseq dataset:
- 30 frames, 18 cameras
- 332,723 Gaussians
- PSNR: 15.39 dB, SSIM: 0.396

+++ datasets/FreeTime_dataset.py
@@ -388,7 +388,10 @@ class FreeTimeParser:
 
             transform = T2 @ T1
         else:
-            transform = np.eye(4)
+            transform = np.eye(4, dtype=np.float32)
+
+        # Ensure camtoworlds is float32
+        camtoworlds = camtoworlds.astype(np.float32)
 
         # Camera paths (folders containing images)
         images_dir = os.path.join(data_dir, "images")
